refactor: update library directory references by removing "rtype_" prefix
