+++ tests/ecs/core/test_entity.cpp
@@ -8,7 +8,7 @@
 #include <gtest/gtest.h>
 #include <unordered_set>
 #include <unordered_map>
-#include "../../../lib/rtype_ecs/src/core/Entity.hpp"
+#include "../../../lib/ecs/src/core/Entity.hpp"
 
 using namespace ECS;
