+++ tests/ecs/core/test_registry_view.cpp
@@ -6,7 +6,7 @@
 */
 
 #include <gtest/gtest.h>
-#include "../../../lib/rtype_ecs/src/core/Registry/Registry.hpp"
+#include "../../../lib/ecs/src/core/Registry/Registry.hpp"
 #include <vector>
 #include <set>
 #include <algorithm>
