+++ tests/ecs/core/test_system_scheduler.cpp
@@ -1,7 +1,7 @@
 #include <gtest/gtest.h>
 
-#include "../../../lib/rtype_ecs/src/system/SystemScheduler.hpp"
-#include "../../../lib/rtype_ecs/src/core/Registry/Registry.hpp"
+#include "../../../lib/ecs/src/system/SystemScheduler.hpp"
+#include "../../../lib/ecs/src/core/Registry/Registry.hpp"
 
 using namespace ECS;
