+++ tests/games/rtype/CMakeLists.txt
@@ -3,7 +3,7 @@ add_executable(test_spawning test_spawning.cpp)
 target_link_libraries(test_spawning PRIVATE
     rtype_game_shared
     rtype_game_server
-    rtype_engine
+    engine
     GTest::gtest_main
 )
 
@@ -28,7 +28,7 @@ add_executable(test_components test_components.cpp)
 
 target_link_libraries(test_components PRIVATE
     rtype_game_shared
-    rtype_engine
+    engine
     GTest::gtest_main
 )
 
@@ -44,7 +44,7 @@ add_executable(test_behaviors test_behaviors.cpp)
 
 target_link_libraries(test_behaviors PRIVATE
     rtype_game_shared
-    rtype_engine
+    engine
     GTest::gtest_main
 )
 
@@ -60,7 +60,7 @@ add_executable(test_behavior_registry test_behavior_registry.cpp)
 
 target_link_libraries(test_behavior_registry PRIVATE
     rtype_game_shared
-    rtype_engine
+    engine
     GTest::gtest_main
 )
 
@@ -77,7 +77,7 @@ add_executable(test_systems test_systems.cpp)
 target_link_libraries(test_systems PRIVATE
     rtype_game_shared
     rtype_game_server
-    rtype_engine
+    engine
     GTest::gtest_main
 )
 
@@ -94,7 +94,7 @@ add_executable(test_collision test_collision.cpp)
 target_link_libraries(test_collision PRIVATE
     rtype_game_shared
     rtype_game_server
-    rtype_engine
+    engine
     GTest::gtest_main
 )
 
@@ -109,13 +109,13 @@ gtest_discover_tests(test_collision)
 add_executable(test_engine test_engine.cpp)
 
 target_link_libraries(test_engine PRIVATE
-    rtype_engine
+    engine
     GTest::gtest_main
 )
 
 target_include_directories(test_engine PRIVATE
     ${CMAKE_SOURCE_DIR}/include
-    ${CMAKE_SOURCE_DIR}/lib/rtype_engine/src
+    ${CMAKE_SOURCE_DIR}/lib/engine/src
 )
 
 gtest_discover_tests(test_engine)
@@ -125,7 +125,7 @@ add_executable(test_projectile_system test_ProjectileSystem.cpp)
 
 target_link_libraries(test_projectile_system PRIVATE
     rtype_game_shared
-    rtype_engine
+    engine
     GTest::gtest_main
 )
 
@@ -141,7 +141,7 @@ add_executable(test_lifetime_system test_LifetimeSystem.cpp)
 
 target_link_libraries(test_lifetime_system PRIVATE
     rtype_game_shared
-    rtype_engine
+    engine
     GTest::gtest_main
 )
 
@@ -157,14 +157,14 @@ add_executable(test_quadtree test_quadtree.cpp)
 
 target_link_libraries(test_quadtree PRIVATE
     rtype_game_shared
-    rtype_engine
+    engine
     GTest::gtest_main
 )
 
 target_include_directories(test_quadtree PRIVATE
     ${CMAKE_SOURCE_DIR}/include
     ${CMAKE_SOURCE_DIR}/src
-    ${CMAKE_SOURCE_DIR}/lib/rtype_ecs/src
+    ${CMAKE_SOURCE_DIR}/lib/ecs/src
 )
 
 gtest_discover_tests(test_quadtree)
@@ -174,14 +174,14 @@ add_executable(test_quadtree_integration test_quadtree_integration.cpp)
 
 target_link_libraries(test_quadtree_integration PRIVATE
     rtype_game_shared
-    rtype_engine
+    engine
     GTest::gtest_main
 )
 
 target_include_directories(test_quadtree_integration PRIVATE
     ${CMAKE_SOURCE_DIR}/include
     ${CMAKE_SOURCE_DIR}/src
-    ${CMAKE_SOURCE_DIR}/lib/rtype_ecs/src
+    ${CMAKE_SOURCE_DIR}/lib/ecs/src
 )
 
 gtest_discover_tests(test_quadtree_integration)
@@ -192,14 +192,14 @@ add_executable(test_projectile_spawner_system test_ProjectileSpawnerSystem.cpp)
 target_link_libraries(test_projectile_spawner_system PRIVATE
     rtype_game_shared
     rtype_game_server
-    rtype_engine
+    engine
     GTest::gtest_main
 )
 
 target_include_directories(test_projectile_spawner_system PRIVATE
     ${CMAKE_SOURCE_DIR}/include
     ${CMAKE_SOURCE_DIR}/src
-    ${CMAKE_SOURCE_DIR}/lib/rtype_ecs/src
+    ${CMAKE_SOURCE_DIR}/lib/ecs/src
 )
 
 gtest_discover_tests(test_projectile_spawner_system)
@@ -209,7 +209,7 @@ add_executable(test_aabb test_aabb.cpp)
 
 target_link_libraries(test_aabb PRIVATE
     rtype_game_shared
-    rtype_engine
+    engine
     GTest::gtest_main
 )
